[BUGFIX LTS] Backport some stable type import locations
The `ContainerProxyMixin` and `RegistryProxyMixin` private types are
used as "intimate API" by `@ember/test-helpers`, which aims to support
both the preview and stable type definitions. Provide both these types
in both their legacy locations (where they were published from until
4.10) and their correct locations. *Only* provide the types, however,
since the runtime location has moved.
