+++ types/preview/@ember/-internals/runtime/lib/mixins/container_proxy.d.ts
@@ -1,0 +1,11 @@
+declare module '@ember/-internals/runtime/lib/mixins/container_proxy' {
+  import { ContainerProxy } from '@ember/-internals/owner';
+  import Mixin from '@ember/object/mixin';
+
+  /**
+   * Given a fullName return a factory manager.
+   */
+  interface ContainerProxyMixin extends ContainerProxy {}
+  const ContainerProxyMixin: Mixin;
+  export default ContainerProxyMixin;
+}
