+++ types/preview/@ember/engine/-private/container-proxy-mixin.d.ts
@@ -1,11 +1,3 @@
 declare module '@ember/engine/-private/container-proxy-mixin' {
-  import { ContainerProxy } from '@ember/-internals/owner';
-  import Mixin from '@ember/object/mixin';
-
-  /**
-   * Given a fullName return a factory manager.
-   */
-  interface ContainerProxyMixin extends ContainerProxy {}
-  const ContainerProxyMixin: Mixin;
-  export default ContainerProxyMixin;
+  export type { default } from '@ember/-internals/runtime/lib/mixins/container_proxy';
 }
